Avoid debugging output charset problems crashing skyped.py
When using irssi in an LC_CTYPE=ISO-8859-1 terminal (thus
/set charset iso-8859-1 for irssi, set charset iso-8859-1 for
bitlbee and no recode_out_default_charset set for irssi) and
skyped.py in a Windows console window (CP1252?), sending a
message with a &eacute; (the character, not the HTML entity)
crashes the server because in its debugging output it tries to
print the UTF-8 equivalent intepreted as CP1252 code points.

I do not think that the "if options.log: ..." code just below the
patch needs the same treatment as the worst that can happen is
that some messages in the log file will be garbled.  Since I do
not use/test that code path, I didn't touch it.